feat: implement concept of packages in diamond.js, equilateral.js,index.js,constants.js, permissions.js, validation.js
