+++ extra/patterns/index.js
@@ -1,0 +1,2 @@
+export {default as diamond } from './Diamond'
+export { default as equilateral } from './Equilateral'
