+++ extra/utils/helpers.js
@@ -1,0 +1,4 @@
+export default function validateEmail(email) {
+    let rex= /^([A-Za-z0-9_\-\.])+\@(successive.tech)$/;
+    return(rex.test(email))
+    }
