+++ extra/utils/index.js
@@ -1,0 +1,2 @@
+export {default as hasPermission} from './permissions';
+export {default as validateUsers} from './validation'
