+++ package.json
@@ -5,19 +5,21 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "validation": "nodemon extra/utils/validation.js",
-    "permissions": "nodemon extra/utils/permissions.js",
-    "pattern:diamond": "nodemon extra/patterns/Diamond.js",
-    "pattern:equilateral": "nodemon extra/patterns/Equilateral.js"
- },
-    "repository": {
-      "type": "git",
-      "url": "git+https://github.com/DPDeepak/express-server.git"
-    },
+    "dev": "babel-node extra --presets env"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/DPDeepak/express-server.git"
+  },
   "author": "Deepak",
   "license": "ISC",
-    "bugs": {
-      "url": "https://github.com/DPDeepak/express-server/issues"
-    },
-    "homepage": "https://github.com/DPDeepak/express-server#readme"
+  "bugs": {
+    "url": "https://github.com/DPDeepak/express-server/issues"
+  },
+  "homepage": "https://github.com/DPDeepak/express-server#readme",
+  "dependencies": {},
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.7.0"
   }
+}
